requirements.txt
python-telegram-bot==20.7
jinja2==3.1.2
pdfkit==1.0.0
